Separated login and account creation
        You navigate to different pages if you're logging in or creating
an account - there's validation if you enter a username that's already
taken on account creation page, or if you enter a username that doesn't
exist on login page. It'll let you know then in top right corner you can
switch to login or create account page depending on which you're already
on.

        If you create a new user, it automatically switches you to the
user page.